backtesting/requirements.txt
pandas>=2.2
python-dotenv>=1.0
openai>=1.50
google-genai>=0.6.0
anthropic>=0.36.0
tenacity>=9.0.0
tqdm>=4.66
